Change NNCF API for training
