+++ README.md
@@ -34,7 +34,38 @@ model = OVAutoModel.from_pretrained(<name_or_path>)
 
 ## NNCF
 
-TBD
+NNCF is used for model training with applying such features like quantization, pruning. To enable NNCF in you training pipeline do the following steps:
+
+1. Import `NNCFAutoConfig`:
+
+```python
+from optimum.intel.nncf import NNCFAutoConfig
+```
+
+> **NOTE**: `NNCFAutoConfig` must be imported before `transformers` to make magic work
+
+2. Initialize a config from `.json` file:
+
+```python
+nncf_config = NNCFAutoConfig.from_json("nncf_bert_config_squad.json")
+```
+
+3. Pass a config to `Trainer` object. In example,
+
+```python
+trainer = QuestionAnsweringTrainer(
+    model=model,
+    args=training_args,
+    train_dataset=train_dataset if training_args.do_train else None,
+    eval_dataset=eval_dataset if training_args.do_eval else None,
+    eval_examples=eval_examples if training_args.do_eval else None,
+    tokenizer=tokenizer,
+    data_collator=data_collator,
+    post_process_function=post_processing_function,
+    compute_metrics=compute_metrics,
+    nncf_config=nncf_config,
+)
+```
 
 ## POT
